check .less files in paths which exist
